+++ NanoXlsx.Core.Test/Colors/ThemeColorTest.cs
@@ -1,4 +1,5 @@
 ﻿using NanoXLSX.Colors;
+using NanoXLSX.Exceptions;
 using NanoXLSX.Themes;
 using Xunit;
 
@@ -13,7 +14,7 @@ namespace NanoXLSX.Core.Test.Colors
             Assert.Equal(Theme.ColorSchemeElement.Dark1, themeColor.ColorValue); // Implicit default value of enum
         }
 
-        [Theory(DisplayName = "Default Constructor Test with arguments")]
+        [Theory(DisplayName = "Default Constructor Test with enum element as argument")]
         [InlineData(Theme.ColorSchemeElement.Dark1)]
         [InlineData(Theme.ColorSchemeElement.Light1)]
         [InlineData(Theme.ColorSchemeElement.Dark2)]
@@ -30,6 +31,35 @@ namespace NanoXLSX.Core.Test.Colors
         {
             ThemeColor themeColor = new ThemeColor(colorSchemeElement);
             Assert.Equal(colorSchemeElement, themeColor.ColorValue);
+        }
+
+        [Theory(DisplayName = "Default Constructor Test with index as argument")]
+        [InlineData(0, Theme.ColorSchemeElement.Dark1)]
+        [InlineData(1, Theme.ColorSchemeElement.Light1)]
+        [InlineData(2, Theme.ColorSchemeElement.Dark2)]
+        [InlineData(3, Theme.ColorSchemeElement.Light2)]
+        [InlineData(4, Theme.ColorSchemeElement.Accent1)]
+        [InlineData(5, Theme.ColorSchemeElement.Accent2)]
+        [InlineData(6, Theme.ColorSchemeElement.Accent3)]
+        [InlineData(7, Theme.ColorSchemeElement.Accent4)]
+        [InlineData(8, Theme.ColorSchemeElement.Accent5)]
+        [InlineData(9, Theme.ColorSchemeElement.Accent6)]
+        [InlineData(10, Theme.ColorSchemeElement.Hyperlink)]
+        [InlineData(11, Theme.ColorSchemeElement.FollowedHyperlink)]
+        public void ConstructorTest3(int givenIndex, Theme.ColorSchemeElement expectedElement)
+        {
+            ThemeColor themeColor = new ThemeColor(givenIndex);
+            Assert.Equal(expectedElement, themeColor.ColorValue);
+        }
+
+        [Theory(DisplayName = "Test of the failing Constructor on invalid values")]
+        [InlineData(-1)]
+        [InlineData(12)]
+        [InlineData(255)]
+        [InlineData(-100)]
+        public void ConstructorFailTest(int value)
+        {
+            Assert.Throws<StyleException>(() => { var color = new ThemeColor(value); });
         }
 
         [Theory(DisplayName = "Test of the StringValue property")]